requirements.txt
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
asgiref==3.7.2
attrs==23.2.0
blinker==1.7.0
boto3==1.34.44
botocore==1.34.44
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
coreapi==2.3.3
coreschema==0.0.4
Django==5.0.2
django-cors-headers==4.3.1
django-environ==0.11.2
django-filter==23.5
django-storages==1.14.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
drf-extra-fields==3.7.0
drf-spectacular==0.27.1
drf-spectacular-sidecar==2024.2.1
drf-writable-nested==0.7.0
Faker==22.7.0
filetype==1.2.0
Flask==3.0.2
Flask-JWT-Extended==4.6.0
grpcio==1.60.1
grpcio-tools==1.60.1
gunicorn==21.2.0
idna==3.6
inflection==0.5.1
itsdangerous==2.1.2
itypes==1.2.0
Jinja2==3.1.3
jmespath==1.0.1
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
Markdown==3.5.2
MarkupSafe==2.1.4
mysqlclient==2.2.4
openapi-codec==1.3.2
packaging==23.2
pillow==10.2.0
protobuf==4.25.2
psycopg2-binary==2.9.9
pycparser==2.21
PyJWT==2.8.0
python-dateutil==2.8.2
pytz==2024.1
PyYAML==6.0.1
referencing==0.33.0
requests==2.31.0
rpds-py==0.17.1
s3transfer==0.10.0
simplejson==3.19.2
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.9.0
uritemplate==4.1.1
urllib3==2.0.7
Werkzeug==3.0.1
